refactor: separate profile image widget
