+++ lib/app/modules/chat/presentation/widgets/pot_user.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:pot_g/app/modules/chat/domain/entities/pot_info_entity.dart';
 import 'package:pot_g/app/modules/chat/domain/entities/pot_user_entity.dart';
+import 'package:pot_g/app/modules/chat/presentation/widgets/pot_profile_image.dart';
 import 'package:pot_g/app/modules/common/presentation/widgets/pot_button.dart';
 import 'package:pot_g/app/values/palette.dart';
 import 'package:pot_g/app/values/text_styles.dart';
-import 'package:pot_g/gen/assets.gen.dart';
 import 'package:pot_g/gen/strings.g.dart';
 
 class PotUser extends StatelessWidget {
@@ -11,36 +12,18 @@ class PotUser extends StatelessWidget {
     super.key,
     required this.user,
     this.onKick,
-    required this.profileIndex,
+    required this.pot,
   });
 
   final PotUserEntity user;
+  final PotInfoEntity pot;
   final VoidCallback? onKick;
-  final int profileIndex;
-
-  Image getImageByIndex(int index) {
-    final images = [
-      Assets.images.jennie.image(),
-      Assets.images.tree.image(),
-      Assets.images.geni.image(),
-      Assets.images.us.image(),
-    ];
-    return images[index % images.length];
-  }
 
   @override
   Widget build(BuildContext context) {
     return Row(
       children: [
-        Container(
-          width: 40,
-          height: 40,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(100),
-            border: Border.all(color: Palette.borderGrey2, width: 0.5),
-          ),
-          child: ClipOval(child: getImageByIndex(profileIndex)),
-        ),
+        PotProfileImage(user: user, pot: pot),
         SizedBox(width: 8),
         Row(
           children: [
